% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeCpp.R
\name{initializeCpp}
\alias{initializeCpp}
\alias{initializeCpp,H5SparseMatrixSeed-method}
\alias{initializeCpp,HDF5ArraySeed-method}
\title{Initialize HDF5-backed matrices.}
\usage{
\S4method{initializeCpp}{H5SparseMatrixSeed}(x, ..., memorize = FALSE)

\S4method{initializeCpp}{HDF5ArraySeed}(x, ..., memorize = FALSE)
}
\arguments{
\item{x}{A \pkg{HDF5Array} seed object.}

\item{...}{Further arguments, ignored.}

\item{memorize}{Logical scalar specifying whether to load the matrix data in \code{x} into memory, if it has not already been loaded.
See \code{\link{checkMemoryCache}} for details.}
}
\value{
An external pointer that can be used in any \pkg{tatami}-compatible function.
}
\description{
Initialize C++ representations of HDF5-backed matrices based on their \pkg{HDF5Array} representations.
}
\examples{
library(HDF5Array)
y <- matrix(runif(1000), ncol=20, nrow=50)
z <- as(y, "HDF5Array")
ptr <- initializeCpp(z)

}
\author{
Aaron Lun
}
